third_party/blink/web_tests/platform/generic/http/tests/websocket/multiple-subprotocols-expected.txt

Test WebSocket subprotocol negotiation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.protocol is ""
Connected
PASS ws.protocol is "superchat"
Received: superchat
PASS receivedMessage is "superchat"
Closed
PASS ws.protocol is "superchat"
PASS closeEvent.wasClean is true
Exited onclose handler
PASS ws.protocol is "superchat"
PASS successfullyParsed is true

TEST COMPLETE
